+++ bundle/config/mutator/initialize_urls_test.go
@@ -25,31 +25,31 @@ func TestInitializeURLs(t *testing.T) {
 			Resources: config.Resources{
 				Jobs: map[string]*resources.Job{
 					"job1": {
-						ID:          "1",
-						JobSettings: jobs.JobSettings{Name: "job1"},
+						BaseResource: resources.BaseResource{ID: "1"},
+						JobSettings:  jobs.JobSettings{Name: "job1"},
 					},
 				},
 				Pipelines: map[string]*resources.Pipeline{
 					"pipeline1": {
-						ID:             "3",
+						BaseResource:   resources.BaseResource{ID: "3"},
 						CreatePipeline: pipelines.CreatePipeline{Name: "pipeline1"},
 					},
 				},
 				Experiments: map[string]*resources.MlflowExperiment{
 					"experiment1": {
-						ID:         "4",
-						Experiment: ml.Experiment{Name: "experiment1"},
+						BaseResource: resources.BaseResource{ID: "4"},
+						Experiment:   ml.Experiment{Name: "experiment1"},
 					},
 				},
 				Models: map[string]*resources.MlflowModel{
 					"model1": {
-						ID:                 "a model uses its name for identifier",
+						BaseResource:       resources.BaseResource{ID: "a model uses its name for identifier"},
 						CreateModelRequest: ml.CreateModelRequest{Name: "a model uses its name for identifier"},
 					},
 				},
 				ModelServingEndpoints: map[string]*resources.ModelServingEndpoint{
 					"servingendpoint1": {
-						ID: "my_serving_endpoint",
+						BaseResource: resources.BaseResource{ID: "my_serving_endpoint"},
 						CreateServingEndpoint: serving.CreateServingEndpoint{
 							Name: "my_serving_endpoint",
 						},
@@ -57,7 +57,7 @@ func TestInitializeURLs(t *testing.T) {
 				},
 				RegisteredModels: map[string]*resources.RegisteredModel{
 					"registeredmodel1": {
-						ID: "8",
+						BaseResource: resources.BaseResource{ID: "8"},
 						CreateRegisteredModelRequest: catalog.CreateRegisteredModelRequest{
 							Name: "my_registered_model",
 						},
@@ -71,7 +71,7 @@ func TestInitializeURLs(t *testing.T) {
 				},
 				Schemas: map[string]*resources.Schema{
 					"schema1": {
-						ID: "catalog.schema",
+						BaseResource: resources.BaseResource{ID: "catalog.schema"},
 						CreateSchema: catalog.CreateSchema{
 							Name: "schema",
 						},
@@ -79,7 +79,7 @@ func TestInitializeURLs(t *testing.T) {
 				},
 				Clusters: map[string]*resources.Cluster{
 					"cluster1": {
-						ID: "1017-103929-vlr7jzcf",
+						BaseResource: resources.BaseResource{ID: "1017-103929-vlr7jzcf"},
 						ClusterSpec: compute.ClusterSpec{
 							ClusterName: "cluster1",
 						},
@@ -87,7 +87,7 @@ func TestInitializeURLs(t *testing.T) {
 				},
 				Dashboards: map[string]*resources.Dashboard{
 					"dashboard1": {
-						ID: "01ef8d56871e1d50ae30ce7375e42478",
+						BaseResource: resources.BaseResource{ID: "01ef8d56871e1d50ae30ce7375e42478"},
 						DashboardConfig: resources.DashboardConfig{
 							Dashboard: dashboards.Dashboard{
 								DisplayName: "My special dashboard",
@@ -128,8 +128,8 @@ func TestInitializeURLsWithoutOrgId(t *testing.T) {
 			Resources: config.Resources{
 				Jobs: map[string]*resources.Job{
 					"job1": {
-						ID:          "1",
-						JobSettings: jobs.JobSettings{Name: "job1"},
+						BaseResource: resources.BaseResource{ID: "1"},
+						JobSettings:  jobs.JobSettings{Name: "job1"},
 					},
 				},
 			},