+++ bundle/deployplan/action.go
@@ -13,7 +13,7 @@ type Action struct {
 
 func (a Action) String() string {
 	typ, _ := strings.CutSuffix(a.Group, "s")
-	return fmt.Sprintf("  %s %s %s", a.ActionType, typ, a.Key)
+	return fmt.Sprintf("  %s %s %s", a.ActionType.StringShort(), typ, a.Key)
 }
 
 // Implements cmdio.Event for cmdio.Log
@@ -21,45 +21,73 @@ func (a Action) IsInplaceSupported() bool {
 	return false
 }
 
-// These enum values are superset to action types defined in the tfjson library.
-// "recreate" maps to the tfjson.Actions.Replace() function.
-// "update" and "update_with_id" maps to tfjson.Actions.Update() and so on. source:
-// https://github.com/hashicorp/terraform-json/blob/0104004301ca8e7046d089cdc2e2db2179d225be/action.go#L14
-type ActionType string
+type ActionType int
 
+// Actions are ordered in increasing severity.
+// If case of several options, action with highest severity wins.
+// Note, Create/Delete are handled explicitly and never compared.
 const (
-	ActionTypeUnset        ActionType = ""
-	ActionTypeNoop         ActionType = "noop"
-	ActionTypeCreate       ActionType = "create"
-	ActionTypeDelete       ActionType = "delete"
-	ActionTypeUpdate       ActionType = "update"
-	ActionTypeUpdateWithID ActionType = "update_with_id"
-	ActionTypeRecreate     ActionType = "recreate"
+	ActionTypeUnset ActionType = iota
+	ActionTypeNoop
+	ActionTypeResize
+	ActionTypeUpdate
+	ActionTypeUpdateWithID
+	ActionTypeCreate
+	ActionTypeRecreate
+	ActionTypeDelete
 )
 
-var ShortName = map[ActionType]ActionType{
-	ActionTypeUpdateWithID: ActionTypeUpdate,
+var actionName = map[ActionType]string{
+	ActionTypeNoop:         "noop",
+	ActionTypeResize:       "resize",
+	ActionTypeUpdate:       "update(id_stable)",
+	ActionTypeUpdateWithID: "update(id_changes)",
+	ActionTypeCreate:       "create",
+	ActionTypeRecreate:     "recreate",
+	ActionTypeDelete:       "delete",
+}
+
+var nameToAction = map[string]ActionType{}
+
+func init() {
+	for k, v := range actionName {
+		if _, ok := nameToAction[v]; ok {
+			panic("duplicate action string: " + v)
+		}
+		nameToAction[v] = k
+	}
+}
+
+func (a ActionType) IsNoop() bool {
+	return a == ActionTypeNoop
 }
 
 func (a ActionType) KeepsID() bool {
 	switch a {
-	case ActionTypeCreate:
-		return false
-	case ActionTypeUpdateWithID:
-		return false
-	case ActionTypeRecreate:
+	case ActionTypeCreate, ActionTypeUpdateWithID, ActionTypeRecreate, ActionTypeDelete:
 		return false
 	default:
 		return true
 	}
 }
 
-func (a ActionType) String() string {
-	shortAction := ShortName[a]
-	if shortAction != "" {
-		return string(shortAction)
+// StringShort short version of action string, without part in parens.
+func (a ActionType) StringShort() string {
+	items := strings.SplitN(actionName[a], "(", 2)
+	return items[0]
+}
+
+// StringFull returns the string representation of the action type.
+func (a ActionType) StringFull() string {
+	return actionName[a]
+}
+
+func ActionTypeFromString(s string) ActionType {
+	actionType, ok := nameToAction[s]
+	if !ok {
+		return ActionTypeUnset
 	}
-	return string(a)
+	return actionType
 }
 
 // Filter returns actions that match the specified action type